FEAT: Move Shared Notes and My Notes toggle button to the Navbar.
